do not return early if other mutation types are observed
"childList" will appear in Firefox when adding spaces to the end of a line. The mutationRecords returned look like this:
```js
[
{type: "characterData", ...}, # Original state
{type: "childList", ...}, # Something we don't care about
{type: "characterData", ...} # This one contains the changes
]
```

Since we exit early we only see the original state and the space we were trying to add is not seen. This replaces that behavior by just filtering out any non `characterData` mutation records.